refactor: reorder tender result step and update imports to TenderResultView across pages
