+++ web/src/hooks/api/useTenderStepStatuses.ts
@@ -143,10 +143,19 @@ export function useTenderStepStatuses(tenderId: number | null, options: UseTende
             status: deriveStatus(!!raData, l11),
         },
         {
+            id: "result",
+            label: "Result",
+            shortLabel: "Result",
+            stepNumber: 10,
+            hasData: !!tenderResult,
+            isLoading: l9,
+            status: deriveStatus(!!tenderResult, l9),
+        },
+        {
             id: "request-extension",
             label: "Request Extension",
             shortLabel: "Extension",
-            stepNumber: 10,
+            stepNumber: 11,
             hasData: !!requestExt,
             isLoading: l13,
             status: deriveStatus(!!requestExt, l13),
@@ -155,19 +164,10 @@ export function useTenderStepStatuses(tenderId: number | null, options: UseTende
             id: "submit-query",
             label: "Submit Query",
             shortLabel: "Query",
-            stepNumber: 11,
+            stepNumber: 12,
             hasData: !!submitQuery,
             isLoading: l14,
             status: deriveStatus(!!submitQuery, l14),
-        },
-        {
-            id: "result",
-            label: "Result",
-            shortLabel: "Result",
-            stepNumber: 12,
-            hasData: !!tenderResult,
-            isLoading: l9,
-            status: deriveStatus(!!tenderResult, l9),
         },
     ];
