+++ web/src/modules/tendering/checklists/DocumentChecklistShowPage.tsx
@@ -9,7 +9,7 @@ import { DocumentChecklistSection } from "@/modules/tendering/checklists/compone
 import { CostingSheetSection } from "@/modules/tendering/costing-sheets/components/CostingSheetView";
 import { RaSection } from "@/modules/tendering/ras/components/RaShow";
 import { TqTenderSection } from "@/modules/tendering/tq-management/components/TqView";
-import { TenderResultSection } from "@/modules/tendering/results/components/TenderResultShow";
+import { TenderResultSection } from "@/modules/tendering/results/components/TenderResultView";
 import { ShowPageLayout } from "@/components/layout/ShowPageLayout";
 import { useTenderStepStatuses } from "@/hooks/api/useTenderStepStatuses";
 import { BidSubmissionSection } from "../bid-submissions/components/BidSubmissionView";
