+++ web/src/modules/tendering/request-extension/RequestExtensionShowPage.tsx
@@ -11,7 +11,7 @@ import { CostingSheetSection } from '@/modules/tendering/costing-sheets/componen
 import { BidSubmissionSection } from '@/modules/tendering/bid-submissions/components/BidSubmissionView';
 import { RaSection } from "@/modules/tendering/ras/components/RaShow";
 import { TqTenderSection } from "@/modules/tendering/tq-management/components/TqView";
-import { TenderResultSection } from "@/modules/tendering/results/components/TenderResultShow";
+import { TenderResultSection } from "@/modules/tendering/results/components/TenderResultView";
 import { Alert, AlertDescription } from '@/components/ui/alert';
 import { AlertCircle } from 'lucide-react';
 import { ShowPageLayout } from "@/components/layout/ShowPageLayout";
